+++ AmongUsMenu.vcxproj.filters
@@ -187,9 +187,6 @@
     <ClCompile Include="user\game.cpp">
       <Filter>user</Filter>
     </ClCompile>
-    <ClCompile Include="user\keyBindsConfig.cpp">
-      <Filter>user</Filter>
-    </ClCompile>
     <ClCompile Include="user\main.cpp">
       <Filter>user</Filter>
     </ClCompile>
@@ -210,6 +207,9 @@
     </ClCompile>
     <ClCompile Include="gui\theme.cpp">
       <Filter>gui</Filter>
+    </ClCompile>
+    <ClCompile Include="user\keybinds.cpp">
+      <Filter>user</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -333,9 +333,6 @@
     <ClInclude Include="rpc\_rpc.h">
       <Filter>rpc</Filter>
     </ClInclude>
-    <ClInclude Include="user\keyBindsConfig.h">
-      <Filter>user</Filter>
-    </ClInclude>
     <ClInclude Include="user\main.h">
       <Filter>user</Filter>
     </ClInclude>
@@ -359,6 +356,9 @@
     </ClInclude>
     <ClInclude Include="hooks\_hooks.h">
       <Filter>hooks</Filter>
+    </ClInclude>
+    <ClInclude Include="user\keybinds.h">
+      <Filter>user</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
